src/requirements.txt

cvxpy==1.2.1
matplotlib==3.5.0
numpy==1.24.2
pyparsing==3.0.7
pyro_ppl==1.8.4
pytest==7.0.0
pytest-cov==4.0.0
setuptools==63.4.1
tqdm==4.64.0
zuko==0.0.5
arviz==0.12.1
hydra-core==1.3.2
omegaconf==2.2.2
chardet==5.1.0
pandas==1.4.0
pyloric==0.0.1
PyYAML==6.0
sbi==0.21.0
scikit_learn==1.2.1
scipy==1.8.0
seaborn==0.11.2
torch==1.13.1+cu116 --extra-index-url https://download.pytorch.org/whl/cu116
functorch==1.13.1
torchdiffeq==0.2.2
torchvision==0.14.1+cu116 --extra-index-url https://download.pytorch.org/whl/cu116
tueplots==0.0.4
jupyter

git+https://github.com/BorealisAI/advertorch.git 
